log endpoint after deploy
